Try putting all the runners in one place
